% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latEq.R
\name{latEq}
\alias{latEq}
\title{Generate a LaTeX Equation from a Linear Model}
\usage{
latEq(
  .lm,
  sub.i = T,
  add.hat = F,
  num.coef = T,
  round.to = 3,
  dep.var = "y",
  ind.var = "x",
  cat.var = "\\\\gamma",
  add = NULL
)
}
\arguments{
\item{.lm}{A linear model object (\code{lm()}).}

\item{sub.i}{Logical. If \code{TRUE}, subscript ``i'' will be added to the dependent variable. Default is \code{TRUE}.}

\item{add.hat}{Logical. If \code{TRUE}, a hat will be added to the dependent variable. Default is \code{FALSE}.}

\item{num.coef}{Logical. If \code{TRUE}, numeric coefficients will be used in the equation. Default is \code{TRUE}.}

\item{round.to}{Numeric. The number of decimal places to round coefficients to. Default is 3.}

\item{dep.var}{Character. The symbol for the dependent variable. Default is ``y''.}

\item{ind.var}{Character. The symbol for the independent variable. Use \code{"NAME"} for variable names. Default is ``x''.}

\item{cat.var}{Character. The symbol for categorical variables. Default is ``\eqn{\gamma}''.}

\item{add}{Character vector specifying additional elements to add to the equation. Default is \code{NULL}.}
}
\value{
A character string representing the LaTeX equation.
}
\description{
This function takes a linear model object and generates a LaTeX equation representation
based on the model's coefficients, terms, and other optional parameters.
}
\details{
The function supports various options for formatting the equation, including adding subscripts, hats,
using numeric coefficients, and including additional elements specified by the \code{add} parameter.

The \code{.var} parameters (dep.var, ind.var, cat.var) can be set to \code{"NAME"} to use the actual variable names
instead of symbols in the equation. For example, setting \code{ind.var = "NAME"} will replace all independent variables with
the actual names of the variables in the output equation.

The \code{add} parameter can take the following values:
\itemize{
  \item "error": Adds the error term to the equation. For example, \eqn{y = mx + b + \epsilon_i}.
  \item "error2": Similar to "error," but includes additional information about the error term, specifying that it follows a normal distribution. For example, \eqn{y = mx + b + \epsilon_i, \epsilon_i \sim \mathcal{N}(0, \sigma^2)}.
  \item "range": Appends a range condition to the equation, indicating that the equation is valid for a range of values of ``i'' (1 to n). For example, \eqn{y = mx + b, 1 \le i \le n}.
  \item "range2": Similar to "range," but the range condition is based on the length of the fitted values of the linear model. For example, \eqn{y = mx + b, 1 \le i \le 30}.
  \item "legend": Generates a legend for the equation, providing information about the symbols used in the equation. This is particularly useful when variable names are used instead of symbols. For example, \eqn{\code{dep.var} =  \beta_1 \cdot \code{ind.var} + \gamma \cdot \code{cat.var} + \epsilon_i}.
}

Note that in order for R to read LaTeX symbols, you must use a double slash, e.g., ``\\\beta''.
}
\examples{
\code{latEq(lm(mpg ~ wt + hp, data = mtcars))}

# Example with custom parameters
\code{latEq(lm(mpg ~ wt + hp, data = mtcars), sub.i = FALSE, add.hat = TRUE, num.coef = FALSE, dep.var = "y", ind.var = "NAME", cat.var = "\\\\beta")}

# Example with error term and range condition
\code{latEq(lm(mpg ~ wt + hp, data = mtcars), add = c("error", "range"))}

# Example with legend and custom variable names
\code{latEq(lm(mpg ~ wt + hp, data = mtcars), add = "legend", dep.var = "Dependent", ind.var = "Independent", cat.var = "Category")}

}
\seealso{
LaTeX Project: \url{https://www.latex-project.org/}
}
\author{
Zach Vig
}
